[spring event] 숫자 형변환 버그 수정

+++ spring-event/src/main/kotlin/io/philo/springevent/app/post/PostDomainEntity.kt
@@ -21,6 +21,6 @@ class PostDomainEntity(
 
     fun updateValueByPercent(percent: Int) {
 
-        this.value += value * (percent / 100)
+        this.value += (value * (percent / 100.0)).toInt()
     }
 }
