+++ spring-event/src/main/kotlin/io/philo/springevent/app/pre/PreDomainEntity.kt
@@ -17,6 +17,6 @@ class PreDomainEntity(
 
     fun updateValueByPercent(percent: Int) {
 
-        this.value += value * (percent / 100)
+        this.value += (value * (percent / 100.0)).toInt()
     }
 }
